requirements.txt
piexif
Pillow
dateparser
tqdm